[cell_fabric] Move shared primitive generation code from pdks to cell_fabric
